ticket carrito

+++ src/controllers/carts.js
@@ -1,9 +1,17 @@
 import { request, response } from "express";
-import { CartsRepository, UsersRepository } from "../repositories/index.js";
+import { CartsRepository, ProductsRepository, UsersRepository } from "../repositories/index.js";
 
 export const getCartById = async (req = request, res = response) => {
   try {
+    const { _id } = req;
     const { cid } = req.params;
+
+    const usuario = await UsersRepository.getUserById(_id);
+
+    if(!usuario) return res.status(400).json({ok: false, msg: 'Usuario no existe!'});
+
+    if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok: false, msg: 'Carrito no válido'});
+
     const carrito = await CartsRepository.getCartById(cid);
 
     if (carrito) return res.json({ carrito });
@@ -33,7 +41,8 @@ export const getCartById = async (req = request, res = response) => {
 
 export const addProductInCart = async (req = request, res = response) => {
   try {
-    const { cid, pid, _id } = req.params;
+    const { _id } = req;
+    const { cid, pid } = req.params;
 
     const usuario = await UsersRepository.getUserById(_id);
 
@@ -42,9 +51,10 @@ export const addProductInCart = async (req = request, res = response) => {
     console.log({ usuario });
     console.log({ cid });
 
-    if(usuario.cart_id === cid){
-        console.log('Son iguales');
-    }
+    if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok: false, msg: 'Carrito no válido'});
+
+    const existeProducto = await ProductsRepository.getProductById(pid);
+    if (!existeProducto) return res.status(400).json({ok: false, msg: 'El producto no existe'});
 
     const carrito = await CartsRepository.addProductInCart(cid, pid);
 
@@ -64,7 +74,17 @@ export const addProductInCart = async (req = request, res = response) => {
 
 export const deleteProductsInCart = async (req = request, res = response) => {
   try {
+    const { _id } = req;
     const { cid, pid } = req.params;
+
+    const usuario = await UsersRepository.getUserById(_id);
+
+    if(!usuario) return res.status(400).json({ok: false, msg: 'Usuario no existe!'});
+    if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok: false, msg: 'Carrito no válido'});
+
+    const existeProducto = await ProductsRepository.getProductById(pid);
+    if (!existeProducto) return res.status(400).json({ok: false, msg: 'El producto no existe'});
+
     const carrito = await CartsRepository.deleteProductsInCart(cid, pid);
     if (!carrito)
       return res.status(404).json({ msg: "No se pudo ejecutar la acción" });
@@ -81,8 +101,17 @@ export const deleteProductsInCart = async (req = request, res = response) => {
 
 export const updateProductsInCart = async (req = request, res = response) => {
   try {
+    const { _id } = req;
     const { cid, pid } = req.params;
     const { quantity } = req.body;
+
+    const usuario = await UsersRepository.getUserById(_id);
+
+    if(!usuario) return res.status(400).json({ok: false, msg: 'Usuario no existe!'});
+    if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok: false, msg: 'Carrito no válido'});
+
+    const existeProducto = await ProductsRepository.getProductById(pid);
+    if (!existeProducto) return res.status(400).json({ok: false, msg: 'El producto no existe'});
 
     if (!quantity || !Number.isInteger(quantity))
       return res
@@ -104,21 +133,21 @@ export const updateProductsInCart = async (req = request, res = response) => {
   }
 };
 
-export const deleteCart = async (req = request, res = response) => {
-  try {
-    const { cid } = req.params;
-    const carrito = await CartsRepository.deleteCart(cid);
-
-    if (!carrito)
-      return res.status(404).json({ msg: "No se pudo ejecutar la acción" });
-    return res.json({ msg: "Producto actualizado", carrito });
-  } catch (error) {
-    console.log("deleteCart -> ", error);
-    return res
-      .status(500)
-      .json({
-        msg: "Se ha producido un error, comuniquese con su administrador",
-      });
-  }
-};
+//export const deleteCart = async (req = request, res = response) => {
+//  try {
+//    const { cid } = req.params;
+//    const carrito = await CartsRepository.deleteCart(cid);
+//
+//    if (!carrito)
+//      return res.status(404).json({ msg: "No se pudo ejecutar la acción" });
+//    return res.json({ msg: "Producto actualizado", carrito });
+//  } catch (error) {
+//    console.log("deleteCart -> ", error);
+//    return res
+//      .status(500)
+//      .json({
+//        msg: "Se ha producido un error, comuniquese con su administrador",
+//      });
+//  }
+//};
